size_t

+++ src/schemas/NDAr/AreaDetectorWriter.cpp
@@ -167,7 +167,7 @@ void AreaDetectorWriter::write(const FileWriter::FlatbufferMessage &Message) {
   FB_Tables::DType Type = NDAr->dataType();
   auto DataPtr = NDAr->pData()->Data();
   auto NrOfElements = std::accumulate(
-      std::begin(DataShape), std::end(DataShape), 1, std::multiplies<>());
+      std::begin(DataShape), std::end(DataShape), size_t(1), std::multiplies<>());
 
   switch (Type) {
   case FB_Tables::DType::Int8:
